t7 b

+++ src/main/java/app/standard/Util.java
@@ -47,6 +47,8 @@ public class Util {
 
             Path filePath = Paths.get(file);
 
+            if(!Files.exists(filePath)) return;
+
             try {
                 Files.delete(filePath);
             } catch (IOException e) {
